+++ PaddleNLP/paddlenlp/datasets/__init__.py
@@ -20,3 +20,5 @@ from .ptb import *
 from .squad import *
 from .translation import *
 from .dureader import *
+from .dureaderrobust import *
+from .dureaderyesno import *
